The client now supports loading the data files from inside the classpath. This means that the data files no longer have to be downloaded separately, and are built into the JAR file.
The translation files are now also included in the JAR files, and I've changed
the loading mechanism. If you run jfreeciv in a non-english locale, you should
see a fair proportion of translated strings working properly.

The jar file is now built with the Main-Class manifest attribute, so that you can
just double click the jar file to run JFreeciv if you have a platform that supports
executable JAR files.